Build & Refactor: add FireStore, add community activity, package activitys
