+++ presentation/src/main/kotlin/team/dahaeng/android/activity/community/CommunityActivity.kt
@@ -1,0 +1,13 @@
+package team.dahaeng.android.activity.community
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import team.dahaeng.android.R
+
+class CommunityActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_community)
+
+    }
+}
